refactor: move Job conversion methods into JobConverter
